chore: refactor code so it doesn't clutter app.js
