+++ src/app.js
@@ -1,40 +1,15 @@
 import Fastify from 'fastify';
-import { getDB, saveDB } from './utils/db/index.js';
-import { v4 } from 'uuid';
+import { general } from './services/general/index.js';
+import { createTodo } from './services/todos/create-todo.js';
 
 const prefix = '/api';
 
 export async function build () {
   const fastify = Fastify({ logger: true });
 
-  fastify.get(prefix, async (request, reply) => {
-    return { success: true };
-  });
+  fastify.get(prefix, general);
 
-  fastify.post(`${prefix}/todo`, async (request, reply) => {
-    const { body } = request;
-    const { title, description, isDone = false } = body;
-    const db = await getDB();
-
-    const id = v4();
-
-    const todo = {
-      title,
-      description,
-      isDone,
-      createdDate: new Date().getTime(),
-      updatedDate: new Date().getTime()
-    };
-
-    db.todos[id] = todo;
-
-    await saveDB(db);
-
-    return {
-      id,
-      ...todo
-    };
-  });
+  fastify.post(`${prefix}/todo`, createTodo);
 
   return fastify;
 }
